+++ tmou-mapa-backend/src/controllers/mod.rs
@@ -5,3 +5,14 @@ pub mod game;
 pub mod message;
 pub mod skip;
 pub mod standings;
+
+use std::vec::Vec;
+
+use crate::db::Item;
+
+// helper function for controllers
+pub fn get_player_level(items: &Vec<Item>) -> i16
+{
+    items.iter().map(|item| item.level).max().unwrap_or(-1)
+}
+
