+++ tmou-mapa-backend/src/tests/discovery_tests.rs
@@ -1,9 +1,3 @@
-#[cfg(test)]
-
-#[allow(unused)]
-use std::env::current_dir;
-#[allow(unused_imports)]
-use std::cmp::Ordering;
 
 #[allow(unused_imports)]
 use chrono::prelude::*;
@@ -11,145 +5,11 @@ use chrono::prelude::*;
 use chrono::{Utc, Duration};
 
 #[allow(unused_imports)]
-use crate::controllers::standings;
-#[allow(unused_imports)]
 use crate::controllers::discovery as dis;
 #[allow(unused_imports)]
 use crate::models::errors::*;
 #[allow(unused_imports)]
 use crate::models::db as db;
-#[allow(unused_imports)]
-use crate::models::api as api;
-#[allow(unused_imports)]
-use crate::osm_models as osm;
-#[allow(unused_imports)]
-use crate::osm_reader::*;
-
-
- 
-
-/////////////////////////////////////////////////////////////////////
-/// Reader
-/////////////////////////////////////////////////////////////////////
-
-#[test]
-fn osm_reader_when_sample_file_given_four_nodes_and_two_ways_emitted()->TmouResult<()>
-{
-    let fname = current_dir()?.join("sample_osm_data.xml");
-    let osm = read_osm_from_file(fname.to_str().unwrap())?;
-    assert_eq!((&osm.nodes).len(), 4);
-    assert_eq!((&osm.ways).len(), 2);
-    Ok(())
-}
-
-#[test]
-fn osm_reader_when_sample_file_given_way1000_contains_nodes123()->TmouResult<()> 
-{
-    let fname = current_dir()?.join("sample_osm_data.xml");
-    let osm = read_osm_from_file(fname.to_str().unwrap())?;
-    let way = osm.ways.get(&1000).unwrap();
-    assert_eq!(way.nodes, vec![1,2,3]);
-    Ok(())
-}
-
-
-#[test]
-fn osm_reader_node_is_correctly_parsed()->TmouResult<()> 
-{
-    let xml = 
-    r#"<osm>
-         <node id="1" lat="2.5" lon="3.5"/>
-         <way id="1000">
-           <nd ref="1"/>
-           <tag k="highway" v="x"/>
-         </way>
-       </osm>"#;
-    let osm = read_osm_from_string(xml)?;
-    assert_eq!(osm.nodes.len(), 1);
-    let node = osm.nodes.get(&1).unwrap();
-    assert_eq!(node.lat, 2.5);
-    assert_eq!(node.lon, 3.5);
-    Ok(())
-}
-
-#[test]
-fn osm_reader_node_is_ignored_when_not_in_way()->TmouResult<()> 
-{
-    let xml = 
-    r#"<osm>
-        <node id="0" lat="2.5" lon="3.5"/>
-        <node id="1" lat="2.5" lon="3.5"/>
-         <way id="1000">
-           <nd ref="1"/>
-           <tag k="highway" v="x"/>
-         </way>
-       </osm>"#;
-    let osm = read_osm_from_string(xml)?;
-    assert_eq!(osm.nodes.len(),1);
-    Ok(())
-}
-
-#[test]
-fn osm_reader_when_malformed_node_supplied_reader_gracefully_continues()->TmouResult<()> 
-{
-    let xml = 
-    r#"<osm>
-         <node malformed="blablabla"/>
-         <node id="0" lat="0" lon="0"/>
-         <way id="1000">
-           <nd ref="0"/>
-           <tag k="highway" v="x"/>
-         </way>
-       </osm>"#;
-    let osm = read_osm_from_string(xml)?;
-    assert_eq!(osm.nodes.len(),1);
-    Ok(())
-}
-
-#[test]
-fn osm_reader_when_malformed_way_supplied_reader_gracefully_continues()->TmouResult<()> 
-{
-    let xml =
-    r#"<osm>
-         <node id="0" lat="0" lon="0"/>
-         <node id="1" lat="0" lon="0"/>
-         <way malformed="blablabla"/>
-         <way id="1000">
-           <nd reg="1"/>
-           <tag k="highway" v="x"/>
-         </way>
-         <way id="1001">
-           <nd ref="1"/>
-         </way>
-         <way id="1002">
-           <nd ref="1"/>
-           <tag k="highway" v="x"/>
-         </way>
-       </osm>"#;
-    let osm = read_osm_from_string(xml)?;
-    assert_eq!(osm.ways.len(),2);
-    let way = osm.ways.get(&1002).unwrap();
-    assert_eq!(way.nodes, vec![1]);
-    Ok(())
-}
-
-#[test]
-fn osm_reader_when_malformed_xml_suplied_fails_gracefully()->TmouResult<()> 
-{
-    let xml =
-    r#"<osm>
-         <node>
-       </osm>"#;
-    match read_osm_from_string(xml)
-    {
-        Ok(_) => panic!("malformed OSM parsing succeeded (should not)"),
-        Err(_) => Ok(())
-    }
-}
-
-/////////////////////////////////////////////////////////////////////
-/// controllers::discovery
-/////////////////////////////////////////////////////////////////////
 
 #[allow(unused)]
 fn item(t: &str, l: i16, n: &str)->db::Item
@@ -158,7 +18,7 @@ fn item(t: &str, l: i16, n: &str)->db::Item
 }
 
 #[test]
-fn controllers::discovery_returns_unchanged_inventory_when_nothing_found()->TmouResult<()> 
+fn discovery_returns_unchanged_inventory_when_nothing_found()->TmouResult<()> 
 {
     // ready for picking level 2:
     let inventory = vec![
@@ -178,7 +38,7 @@ fn controllers::discovery_returns_unchanged_inventory_when_nothing_found()->Tmou
 }
 
 #[test]
-fn controllers::discovery_returns_level_0_puzzles_at_start()->TmouResult<()> 
+fn discovery_returns_level_0_puzzles_at_start()->TmouResult<()> 
 {
     let inventory = Vec::new();
 
@@ -197,7 +57,7 @@ fn controllers::discovery_returns_level_0_puzzles_at_start()->TmouResult<()>
 }
 
 #[test]
-fn controllers::discovery_returns_level_1_puzzles_when_in_level_0()->TmouResult<()> 
+fn discovery_returns_level_1_puzzles_when_in_level_0()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 0, "puzzles-0")];
@@ -218,7 +78,7 @@ fn controllers::discovery_returns_level_1_puzzles_when_in_level_0()->TmouResult<
 }
 
 #[test]
-fn controllers::discovery_returns_level_4_puzzles_when_in_level_4()->TmouResult<()> 
+fn discovery_returns_level_4_puzzles_when_in_level_4()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 4, "puzzles-4a")];
@@ -239,7 +99,7 @@ fn controllers::discovery_returns_level_4_puzzles_when_in_level_4()->TmouResult<
 }
 
 #[test]
-fn controllers::discovery_returns_empty_when_puzzle_level_4_found_and_in_level_5()->TmouResult<()> 
+fn discovery_returns_empty_when_puzzle_level_4_found_and_in_level_5()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 5, "puzzles-5")];
@@ -257,7 +117,7 @@ fn controllers::discovery_returns_empty_when_puzzle_level_4_found_and_in_level_5
 
  
 #[test]
-fn controllers::discovery_returns_nothing_on_level_1_puzzles_at_start()->TmouResult<()> 
+fn discovery_returns_nothing_on_level_1_puzzles_at_start()->TmouResult<()> 
 {
     let inventory = Vec::new();
 
@@ -273,7 +133,7 @@ fn controllers::discovery_returns_nothing_on_level_1_puzzles_at_start()->TmouRes
 }
 
 #[test]
-fn controllers::discovery_returns_badge_level_when_found_at_start()->TmouResult<()> 
+fn discovery_returns_badge_level_when_found_at_start()->TmouResult<()> 
 {
     // ready for a new badge
     let inventory = Vec::new();
@@ -291,7 +151,7 @@ fn controllers::discovery_returns_badge_level_when_found_at_start()->TmouResult<
 }
 
 #[test]
-fn controllers::discovery_returns_badge_level_when_found_when_inventory_somehow_populated()->TmouResult<()> 
+fn discovery_returns_badge_level_when_found_when_inventory_somehow_populated()->TmouResult<()> 
 {
     // ready for a new badge
     let inventory = vec![
@@ -338,7 +198,7 @@ fn controllers::discovery_returns_badge_level_when_found_when_inventory_somehow_
 }
 
 #[test]
-fn controllers::discovery_returns_nothing_when_inventory_already_contains_badge()->TmouResult<()> 
+fn discovery_returns_nothing_when_inventory_already_contains_badge()->TmouResult<()> 
 {
     // ready for a new badge
     let inventory = vec![
@@ -384,7 +244,7 @@ fn controllers::discovery_returns_nothing_when_inventory_already_contains_badge(
 }
 
 #[test]
-fn controllers::discovery_returns_final_badge_when_on_proper_level()->TmouResult<()> 
+fn discovery_returns_final_badge_when_on_proper_level()->TmouResult<()> 
 {
     // ready for a new badge
     let inventory = vec![item("puzzles", 14, "puzzles-14")];
@@ -405,7 +265,7 @@ fn controllers::discovery_returns_final_badge_when_on_proper_level()->TmouResult
 
 
 #[test]
-fn controllers::discovery_returns_nothing_when_on_badge_but_insufficient_level()->TmouResult<()> 
+fn discovery_returns_nothing_when_on_badge_but_insufficient_level()->TmouResult<()> 
 {
     // ready for a new badge
     let inventory = vec![item("puzzles", 13, "puzzles-13")];
@@ -423,7 +283,7 @@ fn controllers::discovery_returns_nothing_when_on_badge_but_insufficient_level()
 
 
 #[test]
-fn controllers::discovery_returns_fakes_on_checkpoint_when_eligible_nothing_owned()->TmouResult<()> 
+fn discovery_returns_fakes_on_checkpoint_when_eligible_nothing_owned()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 0, "puzzles-0")];
@@ -449,7 +309,7 @@ fn controllers::discovery_returns_fakes_on_checkpoint_when_eligible_nothing_owne
 }
 
 #[test]
-fn controllers::discovery_returns_nothing_on_checkpoint_when_not_eligible_nothing_owned()->TmouResult<()> 
+fn discovery_returns_nothing_on_checkpoint_when_not_eligible_nothing_owned()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 0, "puzzles-0")];
@@ -470,7 +330,7 @@ fn controllers::discovery_returns_nothing_on_checkpoint_when_not_eligible_nothin
 }
 
 #[test]
-fn controllers::discovery_returns_subset_on_checkpoint_when_eligible_some_owned()->TmouResult<()> 
+fn discovery_returns_subset_on_checkpoint_when_eligible_some_owned()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 0, "puzzles-0"),
@@ -503,7 +363,7 @@ fn controllers::discovery_returns_subset_on_checkpoint_when_eligible_some_owned(
 }
 
 #[test]
-fn controllers::discovery_returns_nothing_on_checkpoint_when_not_eligible_some_owned()->TmouResult<()> 
+fn discovery_returns_nothing_on_checkpoint_when_not_eligible_some_owned()->TmouResult<()> 
 {
     let inventory = vec![
         item("puzzles", 0, "puzzles-0"),
@@ -640,251 +500,3 @@ fn discover_fake_puzzle_fails_on_checkpoint_when_not_eligible_some_owned()->Tmou
     assert!(!updated_inventory.is_ok());
     Ok(())
 }
-
-////////////////////////////////////////////////////////
-/// controllers::standings
-////////////////////////////////////////////////////////
-
-#[allow(unused)]
-fn team(rank: u16, name: &str, ps: Vec<(u16, bool, u32)>) -> api::TeamStanding
-{
-    let start_puzzles_solved = ps.iter().filter(|(level,_,_)| *level == 1).count() as u16;
-    api::TeamStanding{
-        rank: rank,
-        name: name.to_string(),
-        puzzles: ps.into_iter().map(|(level, dead, min)| (level, api::PuzzleResult {
-            dead: dead,
-            timestamp: chrono::NaiveDate::from_ymd(2020, 11, 6).and_hms(22, min, 00)
-        })).collect(),
-        badge_count: 0,
-        start_puzzles_solved 
-    }
-}
-
-#[test]
-fn is_better_team_returns_alphabetical_for_empty_teams()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", Vec::new());
-    let b = team(0, "Bazinga", Vec::new());
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Less);
-    Ok(())
-}
-
-#[test]
-fn is_better_team_returns_alphabetical_for_equivalent_teams()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", vec![(1,true,1), (2,true,17), (3,true,23), (4,false,51), (5,true,59)]);
-    let b = team(0, "Bazinga", vec![(1,true,1), (2,true,17), (3,true,23), (4,false,51), (5,true,59)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Less);
-    Ok(())
-}
-
-
-#[test]
-fn is_better_team_does_not_prefer_puzzle_visit_over_nothing()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", Vec::new());
-    let b = team(0, "Bazinga", vec![(1,false,0)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Less);
-    Ok(())
-}
-
-#[test]
-fn is_better_team_does_not_prefer_dead()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", vec![(1,false,0)]);
-    let b = team(0, "Bazinga", vec![(1,true,0)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Less);
-    Ok(())
-}
-
-#[test]
-fn is_better_team_prefers_team_with_solved_puzzle()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", vec![(1,true,0), (2,true,0)]);
-    let b = team(0, "Bazinga", vec![(1,false,1), (2,true,1)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Greater);
-    Ok(())
-}
-
-#[test]
-fn is_better_team_prefers_faster_team()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", vec![(1,false,0), (2,false,1)]);
-    let b = team(0, "Bazinga", vec![(1,false,1), (2,false,0)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Greater);
-    Ok(())
-}
-
-#[test]
-fn is_better_team_prefers_faster_team_on_last_solved()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", vec![(1,true,0), (2,true,1), (3,false,1), (4,true,1), (5,false,1), (6,true,1), (7,true,1), (8,false,1)]);
-    let b = team(0, "Bazinga",   vec![(1,true,0), (2,true,1), (3,true,1), (4,false,1), (5,true,1), (6,false,1), (7,true,2), (8,false,2)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Greater);
-    Ok(())
-}
-
-
-#[test]
-fn is_better_team_prefers_team_with_more_solved()->TmouResult<()> 
-{
-    let a = team(0, "Absolutno", vec![(1,true,1), (2,true,1), (3,true,1), (4,false,1), (5,true,1)]);
-    let b = team(0, "Bazinga", vec![(1,false,0), (2,false,5), (3,true,7)]);
-    assert_eq!(controllers::standings::is_better_team(&a, &b), Ordering::Greater);
-    Ok(())
-}
-
-////////////////////////////////////////////////////
-#[allow(unused)]
-fn db_team(name: &str, ps: Vec<(&str, i16, u32)>) -> Vec<db::TeamStandingsItem>
-{
-    match ps.is_empty()
-    {
-        true => vec![db::TeamStandingsItem {
-            team_name: name.to_string(),
-            type_: None,
-            level: None,
-            name: None,
-            description: None,
-            timestamp: None}],
-        false =>     ps.into_iter().map(|(typ, levl, min)| db::TeamStandingsItem {
-            team_name: name.to_string(),
-            type_: Some(typ.to_string()),
-            level: Some(levl),
-            name: Some("item".to_string()),
-            description: Some("item".to_string()),
-            timestamp: Some(chrono::NaiveDate::from_ymd(2020, 11, 6).and_hms(22, min, 00))
-        }).collect()
-    }
-}
-
-
-
-
-#[test]
-fn calculate_teams_controllers::standings_outputs_2_empty_teams_sorted()->TmouResult<()> 
-{
-    let mut a = db_team("Absolutno", Vec::new());
-    let mut b = db_team("Bazinga", Vec::new());
-    a.append(&mut b);
-
-    let expected = vec![
-        team(1, "Absolutno", Vec::new()),
-        team(2, "Bazinga", Vec::new())
-    ];
-
-    let st = controllers::standings::calculate_teams_controllers::standings(a)?;
-    assert_eq!(st.controllers::standings, expected);
-    Ok(())
-}
-
-#[test]
-fn calculate_teams_controllers::standings_outputs_2_complex_teams_sorted()->TmouResult<()> 
-{
-    let mut a = db_team("Absolutno", vec![
-        ("puzzles-fake", 1, 0),
-        ("puzzles-fake", 1, 0),
-        ("puzzles", 1, 5),
-        ("puzzles", 1, 6),
-        ("puzzles", 1, 8),
-        ("puzzles", 1, 7),
-        ("dead", 1, 10),
-        ("puzzles", 2, 20),
-        ("dead", 2, 25),
-        ("puzzles", 3, 30),
-        ("puzzles", 4, 50),
-    ]);
-    let mut b = db_team("Bazinga", vec![
-        ("puzzles-fake", 1, 0),
-        ("puzzles-fake", 1, 0),
-        ("puzzles", 1, 5),
-        ("puzzles", 1, 6),
-        ("puzzles", 2, 20),
-        ("dead", 2, 25),
-        ("puzzles", 3, 30),
-        ("puzzles", 4, 50),
-    ]);
-    a.append(&mut b);
-
-    let mut res_1 = team(1, "Bazinga", vec![
-        (1, false, 5),
-        (2, true, 20),
-        (3, false, 30),
-        (4, false, 50)
-    ]);
-    res_1.start_puzzles_solved = 2;
-
-    let mut res_2 = team(2, "Absolutno", vec![
-        (1, true, 5),
-        (2, true, 20),
-        (3, false, 30),
-        (4, false, 50)
-    ]);
-    res_2.start_puzzles_solved = 4;
-
-
-    let expected = vec![res_1, res_2];
-
-    let st = controllers::standings::calculate_teams_controllers::standings(a)?;
-    assert_eq!(st.controllers::standings, expected);
-    Ok(())
-}
-
-#[test]
-fn calculate_teams_controllers::standings_outputs_correct_badge_counts()->TmouResult<()> 
-{
-    let mut a = db_team("Absolutno", vec![
-        ("puzzles-fake", 1, 0),
-        ("puzzles-fake", 1, 0),
-        ("puzzles", 1, 5),
-        ("puzzles", 1, 6),
-        ("badge", 0, 21),
-        ("puzzles", 1, 7),
-        ("puzzles", 1, 8),
-        ("dead", 1, 10),
-        ("puzzles", 2, 20),
-        ("badge", 0, 29),
-        ("dead", 2, 25),
-        ("puzzles", 3, 30),
-        ("badge", 0, 11),
-        ("puzzles", 4, 50),
-        ("badge", 0, 38),
-        ("badge", 0, 20),
-    ]);
-    let mut b = db_team("Bazinga", vec![
-        ("puzzles-fake", 1, 0),
-        ("badge", 0, 21),
-        ("puzzles-fake", 1, 0),
-        ("puzzles", 1, 5),
-        ("puzzles", 1, 6),
-        ("puzzles", 2, 20),
-        ("dead", 2, 25),
-        ("puzzles", 3, 30),
-        ("puzzles", 4, 50),
-    ]);
-    let mut c = db_team("Corn Flakes", vec![
-        ("puzzles-fake", 1, 0),
-        ("puzzles-fake", 1, 0),
-        ("puzzles", 1, 5),
-        ("puzzles", 1, 6),
-        ("puzzles", 2, 59),
-    ]);
-    let mut d = db_team("Degen a spol", Vec::new());
-    a.append(&mut b);
-    a.append(&mut c);
-    a.append(&mut d);
-
-    let st = controllers::standings::calculate_teams_controllers::standings(a)?;
-    assert_eq!(st.controllers::standings.len(), 4);
-    assert_eq!(st.controllers::standings[0].name, "Bazinga".to_string());
-    assert_eq!(st.controllers::standings[0].badge_count, 1);
-    assert_eq!(st.controllers::standings[1].name, "Absolutno".to_string());
-    assert_eq!(st.controllers::standings[1].badge_count, 5);
-    assert_eq!(st.controllers::standings[2].name, "Corn Flakes".to_string());
-    assert_eq!(st.controllers::standings[2].badge_count, 0);
-    assert_eq!(st.controllers::standings[3].name, "Degen a spol".to_string());
-    assert_eq!(st.controllers::standings[3].badge_count, 0);
-    Ok(())
-}
-
