+++ tmou-mapa-backend/src/tests/mod.rs
@@ -1,2 +1,6 @@
-pub mod tests;
-pub mod rocket_test;
+#[cfg(test)]
+
+pub mod discovery_tests;
+pub mod rocket_tests;
+pub mod skip_tests;
+pub mod standings_tests;
